backend/livekit-agent/requirements.txt
# LiveKit Agents Framework with plugins (v1.5.0 latest as of March 2026)
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.5
livekit-plugins-openai~=1.5

# Multi-modal & Search
aiohttp>=3.9.0
Pillow>=10.0.0

# Environment
python-dotenv>=1.0.0
